+++ edx_toggles/__init__.py
@@ -2,4 +2,4 @@
 Library and utilities for feature toggles.
 """
 
-__version__ = '5.3.0'
+__version__ = '5.4.0'